more breaking up fig_helper, fig_metadata, and show_metadata into figdata modules focused on color mapping, matrix operations, data processing, and specialized modules
