+++ app/dash/components.py
@@ -3,6 +3,7 @@ from dash import dash_table, dcc, html
 import pandas as pd
 
 import app.fig_builder.fig_metadata as fm
+import app.figdata_manager.color_meta as cm
 import app.utils as utils
 
 
@@ -62,8 +63,8 @@ def pandas_df_to_dash_dt(df: pd.DataFrame, num_groups: int) -> dash_table.DataTa
     for i in range(num_groups):
         sdc = {}
         sdc['if'] = dict(filter_query=f"{{cluster}} = {i}")
-        sdc['backgroundColor'] = fm.colors[i % 10]
-        sdc['color'] = fm.text_colors[i % 10]
+        sdc['backgroundColor'] = cm.colors[i % 10]
+        sdc['color'] = cm.text_colors[i % 10]
         style_data_conditional_list.append(sdc)
 
     dash_dt = dash_table.DataTable(
