+++ app/figdata_manager/data_processor.py
@@ -1,28 +1,9 @@
+import numpy as np
 from operator import itemgetter
 import pandas as pd
 
-
-
-def flatten_topics(topics: list, parent_only: bool = False, max_rank: int = None):
-    '''
-    This is the simplest topic flattener and ideally everything should be using it
-    '''
-    out_list = []
-    parents_seen = []
-
-    for i, topic in enumerate(topics):
-        if max_rank and i >= max_rank:
-            break
-        t_bits = topic['topic_key'].split('.')
-        if len(t_bits) <= 1 or t_bits[0] in parents_seen:
-            continue
-        parents_seen.append(t_bits[0])
-        if parent_only:
-            out_list.append(t_bits[0])
-        else:
-            out_list.append(topic['topic_key'])
-
-    return ', '.join(out_list)
+import app.figdata_manager.matrix_operations as mxop
+from app import utils
 
 
 def flatten_and_format_topics_df(df: pd.DataFrame, score_type: str) -> pd.DataFrame:
@@ -32,7 +13,7 @@ def flatten_and_format_topics_df(df: pd.DataFrame, score_type: str) -> pd.DataFr
 
     df = df[['topic_key', 'topic_name', 'raw_score', 'score', 'is_parent', 'tfidf_score']]
     df.rename(columns={'score': 'scaled_score'}, inplace=True)
-    df['parent_topic'] = df['topic_key'].apply(extract_parent)
+    df['parent_topic'] = df['topic_key'].apply(mxop.extract_parent)
     df = df[df['parent_topic'] != df['topic_key']]
     df['total_score'] = df[score_type].sum()
     df.sort_values(score_type, ascending=False, inplace=True)
@@ -40,28 +21,99 @@ def flatten_and_format_topics_df(df: pd.DataFrame, score_type: str) -> pd.DataFr
     return df
 
 
-def extract_parent(topic_key: str):
-    topic_path = topic_key.split('.')
-    return topic_path[0]
+def scale_values(values: list, low: int = 0, high: int = 1) -> list:
+    raw_low = np.min(values)
+    raw_high = np.max(values)
+    raw_range = raw_high - raw_low
+    scaled_range = high - low
+    scaled_values = []
+    for v in values:
+        scaled_v = (v - raw_low) / raw_range * scaled_range + low
+        scaled_values.append(scaled_v)
+    return scaled_values
 
 
-def flatten_df_list_column(col_contents: list, eligible_col_values: list = None, truncate_at: int = None):
+def explode_speaker_topics(speakers: list, topic_type: str, limit_per_speaker: int = None, percent_distrib_list: list = None) -> list: 
     '''
-    Would be a simple list-concat lambda, except that sometimes we want to ignore column values that aren't part of an 'eligible' reference set
+    Expand individual speaker rows containing multiple nested topics into multiple speaker rows each containing one topic
     '''
-    out_list = []
 
-    i = 0
-    for cell_contents in col_contents:
-        if truncate_at and i >= truncate_at:
-            break
-        if eligible_col_values and cell_contents not in eligible_col_values:
-            # skip ineligible values without incrementing truncate_at comparison counter
-            continue  
-        out_list.append(cell_contents)
-        i += 1
+    if not limit_per_speaker:
+        limit_per_speaker = 10
+    topic_field = f'topics_{topic_type}'
+    
+    exploded_speakers = []
+    for spkr in speakers:
+        if topic_field not in spkr or (spkr['word_count'] < 20 and spkr['line_count'] < 3):
+            continue
+        
+        # extract each topic (up to topic_limit) into its own flattened speaker item
+        topic_limit = min(limit_per_speaker, len(spkr[topic_field]))
+        for i in range(topic_limit):
+            flat_spkr = spkr.copy()
+            exploded_speakers.append(flat_spkr)
+            topic = spkr[topic_field][i]
+            flat_spkr['topic_key'] = topic['topic_key']
+            flat_spkr['topic_name'] = topic['topic_name']
+            flat_spkr['rank'] = i+1
+            flat_spkr['dot_size'] = (topic_limit - i) / topic_limit
+            flat_spkr['score'] = topic['score']
+            flat_spkr['raw_score'] = topic['raw_score']
+            # NOTE sad kazoo this was conceived on a false premise, but keeping it in here for now
+            if percent_distrib_list:
+                flat_spkr['scaled_score'] = utils.normalize_score(topic['raw_score'], percent_distrib_list)
+            # extract each topic (up to topic_limit) into its own flattened speaker item
+            del flat_spkr[topic_field]
 
-    return ', '.join(out_list)
+    return exploded_speakers
+
+
+def flatten_speaker_topics(speakers: list, topic_type: str, limit_per_speaker: int = None) -> list: 
+    '''
+    Replace nested speaker topic dicts with concatenated string of topic_keys in speaker rows, dropping speakers with few lines/words in process
+    For reasons I can't recall, I'm being careful to copy each speaker rather than altering the existing speakers
+    '''
+
+    if not limit_per_speaker:
+        limit_per_speaker = 10
+    topic_field = f'topics_{topic_type}'
+    
+    flattened_speakers = []
+    for spkr in speakers:
+        if spkr['word_count'] < 20 and spkr['line_count'] < 3:
+            continue
+
+        flat_spkr = spkr.copy()
+        topics = []
+        if topic_field in spkr:
+            # extract each topic (up to topic_limit) into its own flattened speaker item
+            topic_limit = min(limit_per_speaker, len(spkr[topic_field]))
+            for i in range(topic_limit):
+                topic = spkr[topic_field][i]
+                topics.append(topic['topic_key'])
+
+        flat_spkr[topic_field] = ', '.join(topics)
+        flattened_speakers.append(flat_spkr)
+
+    return flattened_speakers
+
+
+def flatten_and_refine_alt_names(speakers: list, ignore_dupes: bool = False, limit_per_speaker: int = None) -> list:
+    flattened_speakers = []
+    for spkr in speakers:
+        flat_spkr = spkr.copy()
+        alt_names = []
+        if 'alt_names' in spkr:
+            alt_names_limit = min(limit_per_speaker, len(spkr['alt_names']))
+            alt_names = []
+            for i in range(alt_names_limit):
+                if ignore_dupes and spkr['alt_names'][i].upper() == spkr['speaker'].upper():
+                    continue
+                alt_names.append(spkr['alt_names'][i])
+        flat_spkr['aka'] = ', '.join(alt_names)
+        flattened_speakers.append(flat_spkr)
+
+    return flattened_speakers
 
 
 def generate_topic_aggs_from_episode_topics(episode_topic_lists: list, max_rank: int = None, max_parent_repeats: int = None) -> tuple[pd.DataFrame, pd.DataFrame]:
