+++ app/figdata_manager/speaker_topic_scatter_meta.py
