+++ app/pages/character_listing.py
@@ -5,8 +5,8 @@ from datetime import datetime as dt
 import pandas as pd
 
 import app.es.es_read_router as esr
-import app.fig_builder.fig_helper as fh 
-import app.figdata_transformer.color_processor as cp
+import app.figdata_manager.data_processor as dp 
+import app.figdata_manager.color_meta as cm
 import app.pages.components as cmp
 from app.show_metadata import ShowKey
 from app import utils
@@ -127,8 +127,8 @@ def render_speaker_listing_dt(show_key: str):
 
     indexed_speakers_response = esr.fetch_indexed_speakers(ShowKey(show_key), extra_fields='topics_mbti')
     indexed_speakers = indexed_speakers_response['speakers']
-    indexed_speakers = fh.flatten_speaker_topics(indexed_speakers, 'mbti', limit_per_speaker=3) 
-    indexed_speakers = fh.flatten_and_refine_alt_names(indexed_speakers, limit_per_speaker=1) 
+    indexed_speakers = dp.flatten_speaker_topics(indexed_speakers, 'mbti', limit_per_speaker=3) 
+    indexed_speakers = dp.flatten_and_refine_alt_names(indexed_speakers, limit_per_speaker=1) 
     
     speakers_df = pd.DataFrame(indexed_speakers)
 
@@ -148,7 +148,7 @@ def render_speaker_listing_dt(show_key: str):
     speakers_df['actor(s)'].fillna('', inplace=True)
     speakers_df['actor(s)'] = speakers_df['actor(s)'].apply(lambda x: ', '.join(x))
 
-    speaker_colors = cp.generate_speaker_color_discrete_map(show_key, speaker_names)
+    speaker_colors = cm.generate_speaker_color_discrete_map(show_key, speaker_names)
 
     speaker_listing_dt = cmp.pandas_df_to_dash_dt(speakers_df, display_cols, 'character', speaker_names, speaker_colors,
                                                   numeric_precision_overrides={'seasons': 0, 'episodes': 0, 'scenes': 0, 'lines': 0, 'words': 0})
