+++ app/show_metadata.py
@@ -84,93 +84,6 @@ EXTRA_SPEAKER_COLORS = ['PaleVioletRed', 'Tomato', 'Magenta', 'LightGreen', 'Ste
                         'PeachPuff', 'Orchid', 'ForestGreen', 'LightBlue', 'Tan', 'Violet', 'Orange', 'Purple', 'Chocolate',
                         'OrangeRed', 'PapayaWhip', 'DarkSlateBlue', 'MediumSeaGreen', 'DarkOliveGreen', 'PowderBlue', 'Sienna']
 
-TOPIC_COLORS = {
-    'Action': 'DarkGoldenrod',
-    'Comedy': 'Crimson',
-    'Horror': 'MediumSeaGreen',
-    'Drama': 'Fuchsia',
-    'SciFi': 'DeepSkyBlue',
-    'Fantasy': 'Orange',
-    'Thriller': 'MediumBlue',
-    'Crime': 'Maroon',
-    'War': 'Turquoise',
-    'Musical': 'SlateBlue',
-    'Romance': 'Coral',
-    'Western': 'Burlywood',
-    'Historical': 'LightSlateGray',
-    'Sports': 'SpringGreen',
-}
-
-
-# TODO needs to be an algorithm
-BGCOLORS_TO_TEXT_COLORS = {
-    'DarkGoldenrod': 'Black',
-    'Crimson': 'White',
-    'MediumSeaGreen': 'Black',
-    'Fuchsia': 'White',
-    'DeepSkyBlue': 'Black',
-    'Orange': 'Black',
-    'MediumBlue': 'White',
-    'Maroon': 'White',
-    'Turquoise': 'Black',
-    'SlateBlue': 'White',
-    'Coral': 'Black',
-    'Burlywood': 'Black',
-    'LightSlateGray': 'Black',
-    'SpringGreen': 'White',
-    'CornflowerBlue': 'Black',
-    'Chartreuse': 'Black',
-    'Cyan': 'Black',
-    'CadetBlue': 'Black',
-    'DarkCyan': 'White',
-    'Cornsilk': 'Black', 
-    'MediumVioletRed': 'White',
-    'MediumSlateBlue': 'White',
-    'RebeccaPurple': 'White',
-    'Thistle': 'Black',
-    'DarkKhaki': 'White', 
-    'LightSalmon': 'Black',
-    'DeepPink': 'White',
-    'IndianRed': 'Black',
-    'Peru': 'Black',
-    'DodgerBlue': 'Black',
-    'Aquamarine': 'Black',
-    'DarkSeaGreen': 'White',
-    'LawnGreen': 'Black', 
-    'DarkOrange': 'White',
-    'PaleVioletRed': 'Black', 
-    'Tomato': 'Black', 
-    'Magenta': 'Black', 
-    'LightGreen': 'Black', 
-    'SteelBlue': 'White', 
-    'Bisque': 'Black', 
-    'LightCoral': 'Black', 
-    'HotPink': 'Black', 
-    'Gold': 'Black',
-    'BlueViolet': 'White', 
-    'PaleGreen': 'Black', 
-    'Aqua': 'Black', 
-    'RosyBrown': 'Black', 
-    'FireBrick': 'Black', 
-    'Indigo': 'White', 
-    'Olive': 'Black', 
-    'PeachPuff': 'Black',
-    'Orchid': 'White', 
-    'ForestGreen': 'White', 
-    'LightBlue': 'Black', 
-    'Tan': 'Black', 
-    'Violet': 'White', 
-    'Purple': 'White', 
-    'Chocolate': 'White',
-    'OrangeRed': 'White', 
-    'PapayaWhip': 'Black', 
-    'DarkSlateBlue': 'White', 
-    'DarkOliveGreen': 'White', 
-    'PowderBlue': 'Black', 
-    'Sienna': 'Black',
-    'Red': 'White'
-}
-
 
 WIKIPEDIA_DOMAIN = 'https://en.wikipedia.org/wiki/'
